Tidy up codebase
- move source code into `src` dir
- remove unused rollup dependency
- fix package `main` path
- remove unused package `module` path
- consolidate tests: testing features against the main export can happen
in the index test file for now.
