+++ src/index.js
@@ -1,0 +1,3 @@
+const {getPreprocessor} = require('./main')
+
+module.exports = getPreprocessor
