+++ src/main.js
@@ -434,7 +434,25 @@ async function replaceImages(content, options) {
   return processed.content;
 }
 
+/**
+ * @param {Partial<typeof defaults>} options 
+ */
+function getPreprocessor(options = {}) {
+  options = {
+    ...defaults,
+    ...options
+  };
+
+  return {
+    markup: async ({ content }) => ({
+      code: await replaceImages(content, options)
+    })
+  };
+};
+
+
 module.exports = {
   defaults,
-  replaceImages
-};
+  replaceImages,
+  getPreprocessor
+}
